Run separate server tests using mocha/chai directly
