+++ src/server/wtr-config.js
@@ -1,8 +1,6 @@
 import { argv } from 'node:process';
 import { defaultReporter } from '@web/test-runner';
 import { playwrightLauncher } from '@web/test-runner-playwright';
-import { visualDiff } from './visual-diff-plugin.js';
-import { visualDiffReporter } from './visual-diff-reporter.js';
 
 const DEFAULT_PATTERN = type => `./test/**/*.${type}.js`;
 const DEFAULT_VDIFF = false;
@@ -114,10 +112,10 @@ export function createConfig({
 
 	if (vdiff) {
 		config.reporters ??= [ defaultReporter() ];
-		config.reporters.push(visualDiffReporter());
+		//config.reporters.push(visualDiffReporter());
 
 		config.plugins ??= [];
-		config.plugins.push(visualDiff());
+		//config.plugins.push(visualDiff());
 
 		config.groups.push(getVisualDiffGroup(pattern));
 	}
